fix: Correct type conversion for absolute_min_score and relative_threshold_ratio in search_mf_catalog

+++ ask-data/mcp_server/app/tools/search_mf_catalog.py
@@ -28,8 +28,8 @@ def search_mf_catalog(
     max_metrics = int(os.getenv("MF_MAX_METRICS", max_metrics))
     max_dimensions = int(os.getenv("MF_MAX_DIMENSIONS", max_dimensions))
     max_entities = int(os.getenv("MF_MAX_ENTITIES", max_entities))
-    absolute_min_score = int(os.getenv("MF_ABSOLUTE_MIN_SCORE", absolute_min_score))
-    relative_threshold_ratio = int(os.getenv("MF_RELATIVE_THRESHOLD_RATIO", relative_threshold_ratio))
+    absolute_min_score = float(os.getenv("MF_ABSOLUTE_MIN_SCORE", absolute_min_score))
+    relative_threshold_ratio = float(os.getenv("MF_RELATIVE_THRESHOLD_RATIO", relative_threshold_ratio))
 
     cml_token = get_cml_token()
     embed_url = os.getenv("EMBED_RERANK_URL", "").rstrip("/")
